requirements.txt
requests==2.18.1
google_api_python_client==1.6.2
PyYAML==3.12
